requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
requests==2.32.3
youtube-transcript-api